requirements.txt
absl-py==2.1.0
amqp==5.2.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
billiard==4.2.0
celery==5.4.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.1
comm==0.1.2
contourpy==1.2.0
cryptography==41.0.4
cycler==0.12.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
executing==1.2.0
fastjsonschema==2.19.1
flatbuffers==24.3.25
fonttools==4.47.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
gunicorn==21.2.0
h5py==3.12.1
idna==3.4
imbalanced-learn==0.12.0
imblearn==0.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
joblib==1.3.2
joypy==0.2.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.0.3
jupyter_core==5.2.0
keras==3.5.0
kiwisolver==1.4.5
kombu==5.3.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes>=0.3.1,<0.4
mysql-connector-python==8.0.32
namex==0.0.8
nbformat==5.9.2
nest-asyncio==1.5.6
numpy==1.26.3
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.12.1
packaging==23.0
pandas==2.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
# pillow==10.2.0
platformdirs==3.1.1
plotly==5.19.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
pyzmq==25.0.0
redis==5.0.8
referencing==0.33.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.8.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
squarify==0.4.3
stack-data==0.6.2
tenacity==8.2.3
# tensorboard==2.17.1
# tensorboard-data-server==0.7.2
# tensorflow==2.17.0
# tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.3.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.6
Werkzeug==3.0.4
wrapt==1.16.0
yellowbrick==1.5